requirements.txt
asyncstdlib
black
datetime
jsonschema
pylint
pyyaml
telethon
yamllint
tkinter
pillow
pymongo
dnspython
